HELM-133: rename fault-ds to entity-ds
This commit _just_ renames the `fault-ds` (and updates labels
and such) to be `entity-ds` instead.
