+++ src/datasources/entity-ds/partials/config.html
