+++ src/datasources/entity-ds/partials/query.editor.html
