+++ src/datasources/entity-ds/partials/query.html
